comment

+++ app/Helpers.php
@@ -16,10 +16,10 @@ class Helpers
 			$format = (float)number_format($number / 1000, 2) . 'K';
 		} else if ($number < 1000000000) {
 			// Anything less than a billion
-			$format = number_format($number / 1000000, 2) . 'M';
+			$format = (float)number_format($number / 1000000, 2) . 'M';
 		} else {
 			// At least a billion
-			$format = number_format($number / 1000000000, 2) . 'B';
+			$format = (float)number_format($number / 1000000000, 2) . 'B';
 		}
 		$locale = \App::getLocale();
 		if($locale == 'vi'){
